Stricter Debug, User set installation & data dirs
