+++ .vscode/settings.json
@@ -80,6 +80,8 @@
 		"xstddef": "cpp",
 		"xtr1common": "cpp",
 		"xtree": "cpp",
-		"xutility": "cpp"
+		"xutility": "cpp",
+		"coroutine": "cpp",
+		"resumable": "cpp"
 	}
 }
